fixes for onroad signal strength

+++ selfdrive/ui/ui.cc
@@ -401,7 +401,7 @@ static void update_state(UIState *s) {
   scene.started = sm["deviceState"].getDeviceState().getStarted() && scene.ignition;
   if (sm.updated("deviceState")) {
     scene.deviceState = sm["deviceState"].getDeviceState();
-    scene.network_type_string = ui_network_type[scene.deviceState.getNetworkType()];
+    scene.network_type_string = ui_network_type[(int)scene.deviceState.getNetworkType()];
     scene.network_strength = (int)scene.deviceState.getNetworkStrength();
   }
   if (sm.updated("liveLocationKalman")) {
